test: __test__ 폴더에서 각 모듈 테스트 파일로 분리 및 이동
