+++ src/css/index.css
@@ -2,20 +2,25 @@
 @import './modal.css';
 
 * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
 }
 
-ol,ul {
-    list-style: none;
+ol,
+ul {
+  list-style: none;
 }
 
-html, body {
-    height: 100%;
-    -webkit-font-smoothing: antialiased;
+html,
+body {
+  height: 100%;
+  -webkit-font-smoothing: antialiased;
 }
 
-input, button, textarea, select {
-    font: inherit;
+input,
+button,
+textarea,
+select {
+  font: inherit;
 }
