+++ src/js/utils/validate.test.js
@@ -1,5 +1,5 @@
-import validator from '../utils/validator.js';
-import { ERROR_MESSAGE } from '../utils/constants.js';
+import validator from './validator.js';
+import { ERROR_MESSAGE } from './constants.js';
 
 describe('보고 싶은 영상을 검색 했을 때', () => {
   test('입력 없이 버튼을 눌렀다면 error를 throw한다.', () => {
